+++ IsekaiMod/Content/ContentAdder.cs
@@ -66,6 +66,7 @@ namespace IsekaiMod.Content
                 Features.IsekaiProtagonist.CharacterDevelopment.AlphaStrike.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.BetaStrike.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.GammaStrike.Add();
+                Features.IsekaiProtagonist.CharacterDevelopment.Regeneration.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.EnergyImmunitySelection.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.TrainingMontage.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.BodyStrengthening.Add();