+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.0.1] - 2022-11-23
+### Changed
+- Changed `Isekai Drow` to give +2 DEX, +2 INT, + 4 CHA instead of +4 WIS, +4 CHA
 ### Fixed
 - Fixed missing `Drow Poison` ability use limit.
 - Fixed missing speed buff for `Isekai Spriggan` size alteration ability.