+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Nerfed `Gamer` background to give +4 instead of +10.
+
 ## [2.0.1] - 2022-11-23
 ### Changed
 - Changed `Isekai Succubus` to give +2 DEX, +2 INT, +4 CHA, -2 STR instead of +4 DEX, +4 CHA, -2 STR.