+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added mod support for `Spellbook Merge`.
 - Added `Auto Bolster` Overpowered ability.
 - Added `Unlimited Power` Overpowered ability.
+- Added `Regeneration` Character Development feat.
 - Added `Beta Tester` background.
 ### Changed
 - Nerfed `Gamer` background to give +4 instead of +10.