+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Nerfed `Gamer` background to give +4 instead of +10.
 - Changed icon for `True Smite`.
+### Fixed
+- Fixed `Harem Magnet` immunity to be applied after fascinate effect ends.
 
 ## [2.0.1] - 2022-11-23
 ### Changed